Create Step 2: Test A Real Problem
description about resource estimation:
"circuit width" is equal to number of qubits(we don't delete any qubit in this code) which is equal to number of spins of magnetic system
"circuit depth" all RX,RY,RZ rotations that act parallel on each qubit, are counted as one layer together. Each inteaction term like ZiZj requires a combination of CNOTs and rotations so they add about 2 or 3 layers. Trotterization gate, it is consist of different gates. For each Trotter step, the number of layers is repeated.